Move Indenter to utility file
Indenter is likely to be used in other applications, such as code
generation. So move it to a shareable utility file.
