+++ src/ast_dumpr.cpp
@@ -1,47 +1,11 @@
-#include <cstddef>
 #include <iostream>
-#include <string>
 
 #include "ast.hpp"
 #include "ast_dumper.hpp"
 #include "type.hpp"
+#include "util.hpp"
 
 namespace {
-
-class Indenter {
- public:
-  /// @return `size_per_level` * `level` number of `symbol`s.
-  std::string Indent() const {
-    return std::string(size_per_level_ * level_, symbol_);
-  }
-
-  void IncreaseLevel() {
-    if (level_ == 0 /* no limit */ || level_ < max_level_) {
-      ++level_;
-    }
-  }
-
-  void DecreaseLevel() {
-    if (level_) {
-      --level_;
-    }
-  }
-
-  /// @param symbol The symbol used to indent with.
-  /// @param size_per_level The indention size. For example, if the size is `2`,
-  /// each indention level adds 2 `symbol`s.
-  /// @param max_level If equals to `0`, there is no limit.
-  Indenter(char symbol, std::size_t size_per_level, std::size_t max_level = 0)
-      : symbol_{symbol},
-        size_per_level_{size_per_level},
-        max_level_{max_level} {}
-
- private:
-  char symbol_;
-  std::size_t size_per_level_;
-  std::size_t max_level_;
-  std::size_t level_{0};
-};
 
 auto indenter = Indenter{' ', 2, 80};
